add sql query code to endure a Number is returned instead of String for comment_count

+++ models/model.games.js
@@ -18,7 +18,7 @@ exports.updateReviewById = (review_id, inc_votes) => {
 exports.fetchReviewById = (review_id) => {
   return db
     .query(
-      `SELECT reviews.*, COUNT(comments.comment_id) AS comment_count FROM reviews
+      `SELECT reviews.*, COUNT(comments.comment_id)::INT AS comment_count FROM reviews
   JOIN comments ON reviews.review_id = comments.review_id
   WHERE reviews.review_id=$1
   GROUP BY reviews.review_id;`,
